Add files via upload
Python file to read multiple files in a given folder. When the file is read, code locates the root tag and according to that it locates remaining child tags and replaces value corresponding to its key with a specific macro.
Macros are generally placed in a file during deployment of code in different environments.(Ex: the database URL is a macro since it cannot be constant for dev,qa,prod) 
In this approach the master file is not manupulated during deployments.